+++ src/ORMTestInfrastructure/ConfigurationFactory.php
@@ -12,6 +12,7 @@ namespace Webfactory\Doctrine\ORMTestInfrastructure;
 use Doctrine\Common\Annotations\AnnotationReader;
 use Doctrine\Common\Annotations\PsrCachedReader;
 use Doctrine\Common\Annotations\Reader;
+use Doctrine\Deprecations\Deprecation;
 use Doctrine\ORM\Mapping\Driver\AnnotationDriver;
 use Doctrine\ORM\Mapping\Driver\AttributeDriver;
 use Doctrine\ORM\ORMSetup;
@@ -78,6 +79,11 @@ class ConfigurationFactory
         $paths = $this->getDirectoryPathsForClassNames($entityClasses);
 
         if (class_exists(AnnotationDriver::class)) {
+            Deprecation::trigger(
+                'webfactory/doctrine-orm-test-infrastructure',
+                'https://github.com/webfactory/doctrine-orm-test-infrastructure/pull/54',
+                'Since the AnnotationDriver has been removed in ORM 3.0, using annotation-based mapping as the default has been deprecated in 1.16. In 2.0, attribute-based configuration will be the new default. Either upgrade doctrine/orm to >=3.0 or pass a mapping driver when calling `ORMInfrastructure::create*()` methods to make this notice go away.'
+            );
             return new AnnotationDriver($this->getAnnotationReader(), $paths);
         }